requirements.txt
aws-cdk.aws-s3
